Add an endpoint `/backups/running` to show the currently running backup
Because if the backup page is refreshed, the page would lose the `id` of the backup that is running. This endpoint provides a way to get a handle on the currently running backup. When the page is refreshed, (or opened in a new tab), it will start polling again for the running backup status.