Backup Page improvements (#6121)
* Add `BackupProgressStatus` to monitor the state that the backup progress is in.

This info isn't persisted in the DB. The service will prefer to return the runningBackup when possible, to avoid a DB call, and because the progress status isn't saved in the DB. Updated the API to return a `progress_status` field when the backup is in progress.

Other changes
- Fixed the UNDO section of the sql migration 1903001
- Changed some method signatures to return Optional types instead of nulls
- Changed some tests to use assertJ

* Add an endpoint `/backups/running` to show the currently running backup

Because if the backup page is refreshed, the page would lose the `id` of the backup that is running. This endpoint provides a way to get a handle on the currently running backup. When the page is refreshed, (or opened in a new tab), it will start polling again for the running backup status.