Backup Page UI updates (#6129)
Updated Backups Page UI

This required `backupProgressStatus` to be saved in the DB, since we need to know at what stage a backup errored out. Otherwise, the would assume that the error occurred on the last step which was in progress when it polled the server, which may be incorrect.