Add validations around destination directory for SCM materials
In the browser, we can only validation that the path is relative (for POSIX systems anyway). This is ok as
we will display any messages from the server in context at the field.

Currently, this error-to-field mapping from server responses is limited to top-level pipeline config attrs,
as well as materials and stages. It does not yet descend into jobs and tasks. That will wait for another time.