Allow click to pass through on task editor blur when the related element is a button component
When composing the tasks in a job, the user may not hit enter on the last task entered. The task
editor normally auto-saves when the editor is blurred. When the blur event is fired on the task
editor by the user attempting to click the action buttons (e.g., Save and Run Pipeline), the
unsaved task gets saved, but a corresponding click event does not reach the button. Thus, if the
user's intent was to save the pipeline (which is a fair assumption), the user must click the button
again in order for the request to happen. This, while not technically incorrect, is typically
confusing to the user and is most often interpreted as "I have to click twice to get my request
through, so this feels like a bug". This commit aims to remove this perception.

When a blur event is fired, causing the editor to autosave its state, we now detect if there
was a related element in the blur event. The `Event.relatedTarget` property is set when the
action causing the blur was due to a mouse event, such as a click (there are likely other
scenarios that can cause this, but we only care about clicks). This relatedTarget is the
element that was clicked to remove focus from the editor. If this element is a button
component (as determined by a CSS selector), we replay a click action on the button.

Phew, that was a long explanation for the size of the change in this commit.

This commit also does some minor cleanup to use native DOM methods where possible (specifically,
`Element.closest()`).