Ensure that preflight checks set a config origin on the rendered partial configs
This fixes https://github.com/gocd/gocd/issues/6566 where preflight validation would give different
results than in actuality. Specificially, if a pipeline defined in a configrepo referenced another
pipeline in a separate config repo, the validation would fail because GoCD assumed that the preflighted
repo was local instead of remote. (A validation exists such that a pipeline defined locally in cruise-config.xml
cannot depend on a pipeline defined in a config repo).

To fix this, we set an ephemeral config origin (that asserts itself as "remote") during preflight on
the partial config so that it is not miscategorized as being locally defined.