Adding a way around usage data 1 report a day
Need to do this for test drive metrics because we report before the
user can do anything, and won't report again for 24
hours.

We don't know how long they'll leave up the test drive instances, but it
seems reasonable to assume it'll only be for the length of time they're
test driving GoCD.

So, once the CTAs have been clicked, we allow for a single report
regardless of the when the last report was.