This is a combination of 3 commits.
* These 3 commits are required for 19.7.0 quick release which fixes
installer issues.

* The commits are squashed essentially because we can revert this commit
once we cut a 19.7.0 release and on top of it revert the commit which
reverted all the changes post 19.6.0

1. fc61beb9b34540a9068a20aecbe42326483d1e3f
Fix the DIND image

dockerd was being run as the go user, which is not permitted. Instead
run dockerd using a sudo wrapper

2. e380692fed461c56cc893c77dc13cc7a7c2b511d
Adding a way around usage data 1 report a day

Need to do this for test drive metrics because we report before the
user can do anything, and won't report again for 24
hours.

We don't know how long they'll leave up the test drive instances, but it
seems reasonable to assume it'll only be for the length of time they're
test driving GoCD.

So, once the CTAs have been clicked, we allow for a single report
regardless of the when the last report was.

3. 89219ff587accd0968493ceada0b67303a66120f
er, fixing tests