All AgentDAO methodos include code that registers with transaction manager, an afterCommit() hook that is called by transaction manager after commit,
In order to do this, there was quite a bit of boiler-plate code present in each DAO method so,

replaced that into a method that takes a functional interface (Runnable) and registers this runnable with transaction manager.

So now all every methood in DAO need to do is to pass the lambda that neeeds to be executed after comnmit (without any boier-plate code, repeatitive code).