Fix old rails and vm pages UI rendering issue (#6807)
The old rails and vm pages were broken after webpack changes:
	- https://github.com/gocd/gocd/commit/e573c43ce1d82eacbbd77acc43e8f4ec98247c75
	- https://github.com/gocd/gocd/commit/b2e804298581cc7277f73e770a83605f9c9fb223

Problem:
With the latest webpack, the normalize and typography css is included
on all GoCD pages (because of header/footer component is included on all the GoCD pages)
causing the component-specific global styles to be imported on all the old GoCD pages.

Fix:
Do not include normalize and typography css in older rails and vm pages.
- Extract global css into a separate stylesheet of its own.
- Define a new js-entrypoint for the global css, which requires global css.
- Load this global-css-entrypoint on component layout.