Fixed scrolling issue for `Anchor`
Scroll for element is called from `view` method of `Anchor` when the element is attached to dom and is not already scrolled.

When `view` method is called for the first time, scrolling does not take place as the element is not attached to dom. So the scroll actually takes place in re-render.

To fix this, `scrollToEl` is called from both `onCreate`(callback which gets triggered after the element is attached to dom) and `view` methods so that scroll will happen immediately after the element is attached to dom and on re-render.