Fix inconsistent behaviour for delete group
If the group contains only config-repo pipelines, the DeletePipelineCommand would receive any empty group as we send configForEdit to the same which does not contains the PartialConfigs. Hence, delete is successful which leads to errors later on.

Moved the empty check to canContinue wherein we get the group to be deleted which is via go config service