Don't add GO_AGENT_RESOURCES on null resources
It may happen (and we learned from running the fast-unit tests) that the ResourceConfigs on the agent are null. That means basically there are no defined resources on that agent. Instead of putting an "empty string" it's better if we only define this env. var. only if there's something to populate it. It actually would server users better as they can test for existence of certain variables instead of getting empty strings (and in case of non-existence they can default to a pre-set variable, like in Ansible's default () method).