Adding GO_AGENT_RESOURCES as environment variable (#7589)
* Adding GO_AGENT_RESOURCES as environment variable

The issue fixes upstream GoCD #7575 issue by serializing the agent resources to an environment variable.

* Moved logic for GO_AGENT_RESOURCES in the BuildAssignmentService and enhanced the existing scaffolding tests around secrets/environment with the check for the agent resources environment variable;

* Don't add GO_AGENT_RESOURCES on null resources

It may happen (and we learned from running the fast-unit tests) that the ResourceConfigs on the agent are null. That means basically there are no defined resources on that agent. Instead of putting an "empty string" it's better if we only define this env. var. only if there's something to populate it. It actually would server users better as they can test for existence of certain variables instead of getting empty strings (and in case of non-existence they can default to a pre-set variable, like in Ansible's default () method).