<KeyValEditor/> component exposes a UI to modify a set of encryptable key-value entries.
The typical model to back this component is a set of configuration properties.

This editor supports the new serialization option to represent plain-text values that
a client wishes to encrypt:

```json
{
  "key": "some.property",
  "value": "I will be encrypted by the server upon completion",
  "secure": true
}
```

The intent is to use this in the config_repo modal so users can define variables to be
used in pipeline definition files.