Fix stale data issue with draggable table component.
Problem:
	dragStart, dragOver and dragEnd interface methods were defined
on the state as part of the oninit method of the table component. These
methods uses attributes passed to the table component.

	But as the oninit method is called only once during the time of
table component creation, all the interface method had references to
the attributes which were passed during the table creation. Afterwards,
even when the data was updated, these methods used to operate on the
stale data.

Fix:
	Do not define the methods which uses the component attributes as
part of the component state. Always operate on the vnode object, as the
child references of the vnode are updated, but not the actual vnode.