Fix `<DependencyFields/>` stage dropdown population on modal load
The stages dropdown initially starts with no data because autocomplete
suggestions are async fetched via AJAX, and mithril initially draws the
dropdown with no options.

Eventually, the request finishes and populates the autocomplete cache.
However, the mapped/calculated `Stream` that generates the stage select
field options caches its old value. It appears that one must try to set
data on the mapped `Stream` in order to recalculate the result. Since
`pipeline` is the upstream `Stream` and doesn't change without user
intervention, the calculated value for the stages options `Stream` shows
stale data.

The easy fix is to just not use a mapped `Stream`. Instead, make it a
simple arrow function to calculate the stages from the pipeline name every
time.