Exec task argument fixes
* Allow users to specify whitespaces (do not trim whitespaces)
* Save user specified arguments.

Description:
	User specified arguments were not saved due to invalid
json submitted to the payload. When both 'args' and 'arguments'
are submitted to the server, server will only parse 'args' field.
	As the multiline arguments are submitted to the server
as 'arguments' and empty value for 'args' was submitted to the
server, server parsed only 'args' and saved empty arguments.

Fix: Do not submit args when not specified.