Finish ConfigRepo v3.0 extension API and add capabilities
v3.0 was never published/documented because it was born from the
now-defunct test drive enhancements. It was waiting on the release
of the new PaC page (which has been since merged).

As luck would have it, I could then freely add the correct capabilities
for this version, specifically flags to indicate support for:

  - enumerating config files stored in a given repo (i.e., get-config-files
    request)
  - allowing user properties to be referenced from a config file

As part of branch support, the latter flag is used to conditionally
display the user-defined property entry widget in the configrepo
create/update modal.