We must also factor ConfigOrigin into the PartialConfig equivalence condition
Testing equality of ConfigOrgin (and thereby updating the CruiseConfig
based on it) is important to pipeline scheduling logic. Once a material
receives a new revision, its corresponding pipeline(s) will only be
triggered if the revision matches the config origin of the associated
PipelineConfig instances that define the very pipelines. Thus, we must
continue to update the CruiseConfig if the ConfigOrigin of the
PartialConfig differs, even if the resulting structure has not changed.