Stop optimistically removing config repo merge/validation errors from ServerHealthState
The errors in ServerHealthState should only be removed when a) the
partial is marked as valid and b) when destroying a configrepo config.

This commit has a lot of noise from renaming classes and methods, but is
logically very short -- effictively a one line change.

The heart of this fix is removing the line `serverHealthService.removeByScope(...)`
from `GoConfigRepoConfigDataSource#onCheckoutComplete()`.
`CachedGoPartials#markAsValid()` already removes the error from `ServerHealthState`.

This should fix the regression failures around ConfigRepo validations.