Git materials can checkout arbitrary refs
  - GoCD's `git` layer knows how to handle refspecs provided in the branch field.
  - Refspecs _**must NOT**_ contain wildcards; they are not meaningful to how GoCD
    will use materials (i.e., updating a local agent workspace).
  - The main motivation behind this is to allow GoCD to support pull request
    refs as a pipeline material for the branch support work.

Implementation notes
--------------------

`RefSpecHelper` handles detection and parsing of a given branch or refspec. Mainly,
this utility class extracts source and destination components of the refspec and
builds sensible values for local branch names and remote branch names. Also notably,
it infers a full destination ref from a shorthand refspec. The logic for this is
straightforward:

1. When given a shorthand refspec in the form: `refs/namespace/foo:bar`
    - the refspec destination is expanded under the `origin` remote as
      `refs/namespace/foo:refs/remotes/origin/bar`.

2. When given a full refspec in the form: `refs/namespace/foo:refs/anything/bar`
    - the refspec is interpreted as-is.

`GitCommand#clone()` and `GitCommand#cloneWithNoCheckout()` have been adapted to
accommodate refspecs; when a refspec is detected, the clone behaviors configure
the `git-fetch` configurations with the refspec instead of the implicit/default
`+refs/heads/*:refs/remotes/origin/*` so that future `git fetch` operations update
the correct refs. This ensures that flyweights and agent sandboxes receive new
commits. The local branch is also explicitly created before switching the local
repo `HEAD` pointer.

`GitMaterial` is also refspec aware and knows how to report the local branch name
from the given refspec. Additionally, `GitMaterial#isRepositoryChanged()` performs
additional validation on the fetch configuration before reusing (or re-cloning) a
flyweight or agent sandbox for a given material.

`GitCommand` and `GitMaterial` delegate to `RefSpecHelper` for the detection and
interpretation of refspecs.

There were also some convenience methods added to `SCMCommand` and `CommandLine`
in addition to miscellaneous code formatting, cleanup, and reorganization.