`PartialConfig`s should retain their records of `Rules` violations
Context
-------

This fixes an issue, discovered by @kritika-singh3, wherein a
`PartialConfig` may be merged despite violating rules. This merging
happens when 2 conditions are true:

  1. `PartialConfig` in question is otherwise structurally valid,
     only failing to meet the rule criteria.
  2. There is a second `PartialConfig` that is also in an error
     state, owned by another `ConfigRepoConfig`.

Fixing the errors in the other `PartialConfig` will inadvertently
cause the first PartialConfig (which violates rules but is elsewise
valid) to merge. In fact, even the error disappears, despite no
permissive rules present.

This has to do with the fact that rule violations are checked only
once for a partial before it is added as a merge candidate AND that
we do not persist these violations over time.

Now, the config merge and save process is deep, complicated, and
aneurysm-inducing, so I will highlight some key events involved in
this issue with the hopes that one can understand enough to get the
gist.

The First Time Around
---------------------

When we merge partials, GoCD keeps cached copies of partials. There
are two cache lines for partials: the "last (i.e., previously) valid"
and the "last attempted" (actually called "lastKnown". The very first
thing that happens to our candidate partials is that they are cached
to `lastKnown`, and if after all merging and validation they make it
through the whole process successfully, they are promoted to the
`lastValid` cache. That is, on success: `lastValid == lastKnown`.

The actual merge process happens via "update commands" to the main
config.  In this case, the `PartialConfigUpdateCommand` validates
config domain object references against the `Rules` of the owning
`ConfigRepoConfig`. This validation operates on a *deep-clone* of
the original `PartialConfig`, and is passed further down the merge
and save process. It's worth noting that because this happens on a
deep-clone, the cached partials do not retain the rule violations.

Further down the line, the partial, *along with every other partial*,
is merged into a deep-clone of the main config during a "preprocess"
step. This merged clone is then validated by a tree traverser. At
the end of the validation step, all errors are collected from the
nested hierarchy, including those rule violations that were attached
by `PartialConfigUpdateCommand#update()`. The validation fails, and
the partials are not merged, and the world is seemingly happy.

But alas, it is not.

The Second Time Around
----------------------

What happens when we fix the other `PartialConfig`? We start the merge
process all over again. Partials are collected from the caches along
to be merge candidates, while the newly fixed partial is passed to
`PartialConfigUpdateCommand#update()`. This is the only partial that
will have its rules evaluated; the other candidates have not been
updated, so GoCD does not wrap them in `PartialConfigUpdateCommand`s.
The fixed partial is passed down the line.

Fast-forward to the the preprocess and validation steps. Again, partials
are merged into a cloned copy for validation. The key here is that the
*other* partials, including our original `PartialConfig` that failed
rule validation in the previous merge were pulled from the cache and
have not had rules evaluated again. The cached copies retained no rule
violations because all violations were attached to copies that were
later discarded. As such, when structural validations run, previous
rule violations are not detected, and the merge AND validation succeeds.
The newly merged config becomes the active config, and all partials
are promoted to `lastValid` -- a process that also removes errors from
server health messages.

The end result? The partial, which should not have been merged, was
merged and the user-visible errors have been wiped away.

Oops!

Which Brings Us to the Fix
--------------------------

Yes, this is probably the longest you've read before getting to the
details of this commit. Sorry.

This fix does the following to address the aforementioned issue:

  - Validate rules on the original copies of the partials so that any
    violations survive to the next generation of merge attempts. This
    makes visible the rule violations of non-primary partials (i.e.,
    the partials that are not the subject of the update command) so
    as to be considered before approving a config merge.
  - Move the rule validation logic to live under `PartialConfig` so
    that it is easily shared and callable in case we need to do this
    elsewhere.
  - Ensure server health error removal and `markAsValid()` promotion
    happens only downstream of an update command.