Fix edge case for git refspec clone
As git does not allow you to fetch into/update the current branch, there
exists an edge case when the default branch is the destination of the
provided refspec.

Let's say the default branch is `master`, as it is in nearly every git
repository.

If the refspec is `refs/heads/master:master` (the src doesn't matter --
just the dst), `GitCommand#cloneWithNoCheckout()` will fail on the
`git branch -f master origin/master` step because we are trying to
update the current branch (as `git clone` automatically selects the
default before we switch).

To fix this, we enter a detached head state at the current SHA. This
makes `git branch -f <local> <remote>` truly idempotent.