Solve issue around missing SCMs coming from config repo
 - during converison of configs from config repo to one stored by GoCD, we check if the said scm is already present or not. If former, we set the same as material config and ignore the one defined.
The search for presense was on all the configs including the ones present in the partials.
On the first parse, the SCM would not be found - hence the same would be added to the partial. On save, this will get integrated with all the scms.
However on the next parse it would be found - hence the same will not be added to the partial. On save, this will get integrated with the ones defined in the scm, deleting the older partial - effectively removing the scm
Fix: if the scm is already present - check the origin. If the origin is the current config repo - add it to the partial