Do not purge the config repo defined structures cache on polled fetch
This fixes a UI flicker when a config repo widget is expanded. Previously, the
defined pipeline structure graph would quickly blank out, then reappear each time
GoCD fetched config repo data on interval. This was caused by mithril redrawing
an empty pipeline structures data cache (resulting in an empty graph with a loading
message) followed only milliseconds later by a redraw when data was full again on
fetch response. On most connections, even those that are relatively slow, this looks
like a UI "glitch" because it happens so quickly.

Even though the caches were designed to not clear their contents on `invalidate()`
precisely to address this issue, it was happening anyway because a new cache object
was recreated on each fetch. This commit fixes this by persisting object caches over
intervals for reuse. Data is still updated when it changes, but no longer flickers
because the widget continues to display stale data while waiting for the API response
until fresh data overwrites it upon response completion.