Abstract ConfigRepos fetch behind an ObjectCache
This consolidates the fetching, resolution, and etag management behind a single
object. This lets us separate concerns with regards to fetching config repos and
fetching permissions. As a side effect, we can now handle 304s in refreshConfigRepos()
because of this modularization (previously, we weren't passing the etag along during
the poll, and it would have been clumsy to handle it the way it used to be).