Introduce webhook URL suggestion helper component
This component generates configrepo webhook URLs for chosen providers.
It is shown on the config repo page for config repos where `autoUpdate`
is set to `false`.

This relies on a base-URL building utility which first consults any
configured site URL or secure site URL, falling back to guessing the
base URL from the `window.location` if neither is configured.

The component also informs the user of the provider-specific events to
configure the webhook.

This also introduces a new click-to-copy component that can be composed
more flexibly in that it does not rely on an input field. It will probably
replace <CopyField/>.