Validate agent UUID during remoting requests to prevent clients from acting on behalf of other agents
Prior to this change, it was possible for an authenticated agent to act on behalf of another agent.
This is a security issue because if one agent is compromised, an attacker could request work or perform
other duties intended for other agents. This was possible because the UUID to perform work was sourced
from the deserialized object stream and not verified to be the same UUID that was permitted by the
authentication filter. Essentially, an authenticated client could request work for any arbitrary agent.
This would give access to decrypted secrets for a build.

This commit introduces a custom `HttpInvokerServiceExporter` that validates the UUID from the
`X-Agent-GUID` header matches the UUID from the deserialized `AgentRuntimeInfo` and `AgentIdentifier`
objects and returns an appropriate error code when this validation fails.