Align the JRE-on-Alpine pre-requisites with the latest available compatible versions
This was earlier taken from AdoptOpenJDK (now Eclipse Temurin), however after JDK 15 they abandoned this approach (for a mixture of reliability and security reasons - glibc-on-Alpine is not supported by Alpine team or considered good practice) and instead moved to native Alpine builds of JDK 16+ which are supported within OpenJDK upstream.

However for us, mure musl/alpine JRE builds don't work for us due to use of the Tanuki Wrapper. The wrapper currently needs glibc and does not seem to work against lightweight compatibility shims such as gcompat or libc6-compat :-(

If there is a way to get the wrapper to use glibc, but have the JREs still use musl, that would be a better path forward, however that is probably not possible due to the way the wrapper processes work and inject native code.