Corrects usage of filesize library to the previous behaviour
`filesize` changed the default to IEC standard for base 2 units (KiB, MiB, GiB etc), and default to base 10, which would use SI units of display. We seem to be using base 2 with KB/MB/GB etc (JEDEC standard?) in other places such as the server backup size, and Commons IO is written that way for some server-side stuff, so not changing at this stage, even though it's probably not desirable.

There is an argument that we'd be better to use base 10 (IEC units) rather than base 2, however rather than change this behaviour this just keeps the original behaviour. (ref policies such as https://wiki.ubuntu.com/UnitsPolicy )

https://github.com/avoidwork/filesize.js/compare/7.0.0...8.0.0