Address issues with flaky integration tests when run together in certain combinations and orders
It seems when tests use/involve the fixtures for pipelines they can leave the server configuration/context in a broken state which is not reset correctly for subsequent tests. In particular `PipelineScheduleServiceTest` fails repeatably if run after `JobAssignmentIntegrationTest` or `AgentServiceIntegrationTest`, as do some other tests which expect jobs to be run (or not run) on all agents.

I don't fully understand what is going on here, and did a lot of trial+error to fix, but it seems to relate to not resetting some combination of the server configuration and database configuration to the target state. Ordering appears to also be relevant - reset `DatabaseAccessHelper` first, then `GoConfigFileHelper`. Also changed the fixtures to use `GoConfigFileHelper.onTearDown` for consistency/clarity. It's confusing to have two different approaches used in various test teardowns.

Sometimes a re-run will fix the issue due to indeterminism in the way Gradle orders tests, however other times it's not possible to fix in this way due to the particular combination of tests being run.