Exclude the transitive `slf4j-api` dependency from the `providedAtPackageTime` dependency
Otherwise, if the *version* of slf4j-api transitively referred to by logback is *exactly* the same as that referred to in `implementation` dependency, the whole slf4j API jar is excluded from the lib folder in the server war. That doesn't seem intentional, even though the jar is also included in the launcher, and logback is ONLY present in the launcher.