Avoid full re-installs of Yarn dependencies/Yarn re-runs when using `YarnInstallTask` and `YarnRunTask`
In #8557 the working directory was redeclared as `InputDirectory`. The problem is that using this type causes Gradle to observe the contents of this directory and assess whether it has changed between runs in order to decide whether the task is stale and needs to be re-run. https://docs.gradle.org/current/javadoc/org/gradle/api/tasks/InputDirectory.html

Since this folder has many other files other than `node_modules`, `package.json`, `yarn.lock` etc, it basically always considers the task `outOfDate` which currently causes the code as written to wipe `node_modules` and do an unnecessary full reinstall and re-link, which is **very** slow.

Changing these types back to regular input strings will cause it to only consider them out of date if the path itself changes, which reverts to the behaviour before #8557, while still being Gradle 7 compliant in terms of annotating inputs