Add workaround with problem with use of refinements with JRuby <= 9.2.14.0 and ruby-i18n >= 1.8.8
Not sure if this should stay. There is a secondary problem with JRuby `9.2.9.0` with use of refinements in
certain modes noted at https://github.com/jruby/jruby/issues/6038 This was causing `admin/configuration` tests to fail when using `time_ago_in_words` and then crashing with a nonsensical error rather than the root cause error from the JRuby/i18n problem.

This appears to be fixed in JRuby `9.2.10.0` however this version of JRuby involves RubyGems 3 which has some scripting issues on Windows currently. Groan. Believe refinements in this way shouldn't really be used anywhere we are relying on based on a global search.