Make punctuationRegex static within `IgnoredFiles`
For some reason it was not static before, which is causing it to be serialized/deserialized with Gson. It's non-trivial to serialize a regex, and this does not seem required at all, as it's immutable. Furthermore, on Java 16/17 this requires special reflective access to `java.util.regex` within the `java.util` module on both server and agent, which seems unnecessary.