Use consistent approach to clone cruise config
There are optimizations in the way (most of) configuration is actually cloned in the `GoConfigCloner`. To avoid weird issues in tests, this tries to use the same cloner configuration consistently, at least when cloning the entire `CruiseConfig`. This isn't bullet-proof, as there are actually a number of cloner instances used in different parts of the code that seem to clone different sub-parts of the config, so some of the changes here may not be perfect. Intent is to ensure tests use the same optimizations and "nulling of cached values" as the production code.