Use a subfolder for agent JARs when running server locally
Copying these files into the root of the server project causes Gradle to think we are modifying the input to other tasks without declaring dependencies. On Gradle 7 it disables optimizations (seems bad) and on Gradle 8 it will probably be broken. Using a subfolder seems to work fine; this change uses the same subfolder as used by the server when running in a "real" environment and loading from the classpath.