Include transitive dependencies from `packagingOnly` configuration by default
Generally we need these, for example when pulling bouncy castle PKIX which has a transitive 'util' dependency which is required. The only case we are excluding these is when pulling the `base` dependencies for packaging. That seems a bit crude, but seems to relate to these being simple lib dependencies which are included at top-level on the server side alongside Jetty etc.