Remove the old Ruby/Rails preferences/notifications pages
This was replaced by a new SPA in #8282 with a toggle to reinstate the old page/route. Cannot find any complaints that indicate we might need to retain the toggle/old code. This also seems to be the last Ruby/Rails SPA to be replaced, allowing some removal of cruft.