Ensure Gradle itself can run on Java 17 by working around issue with use of Grolifant and DownloaderTask.groovy on Java 17
There appears to be issues with use of Groovy-generated dynamic proxies for interfaces (with default methods) on Groovy 3.0.9. This workaround opens access to the module system, so that we can then open access to the dynamic proxies. Intend to follow up with Grolifant folks as well as Groovy folks after a bit more clarity on possible backport, at https://issues.apache.org/jira/browse/GROOVY-10145

This currently needs to be addressed since we have custom scripting to download Selenium Drivers for Chrome and Firefox for use in (some) UI tests. May want to evaluate pre-existing automation to handle this later.