Disable GPG signing on the config repository
There is no way to support this currently, as we do not configure JGit for GPG. If a user/host has GPG signing enabled at a global git level, even with GPG configured, this is insufficient to allow JGit to be able to sign commits. Currently we are better to force disable this on the config repository, overriding any global user config that might be inherited.