Downgrade Tanuki wrapper to 3.5.46 (as per last release)
Wrapper versions 3.5.47+ seem to have some funny behaviour in some cases where they spin in a tight CPU loop. Only observed this on build.gocd.org where there is a custom GoCD Server docker image, and supervisord is wrapping launch of the docker entrypoint and the tanuki wrapper, however it is replicable in that specific scenario on both 3.5.47 and 3.5.48. The problem goes away with a downgrade.

There may be a config oriented fix, but not sure right now. https://wrapper.tanukisoftware.com/doc/english/release-notes.html#3.5.47