Fixes a memory leak in JMSMessageListenerAdapterTest
This class starts a thread in the background and doesn't provide a way to interrupt it, except in the real world when the consumer is closed and returns null messages. That doesn't happen with the mocks here unless we tell it to, by resetting the mocks.

Without this happening, the invocation counts on the mocks just keep going on forever in the loop, after a couple of minutes there are 500,000 entries recorded by the mocks somehow :-)

There might also be issues with the daemon thread here for other tests, probably mitigated in integration tests by running each test class in its own JUnit process.