Refactor the TFS command's NestedJarClassLoader to not require internal access
Use of setAccessible is blocked by default on Java 17. The old implementation was directly referring to the parent classLoader in a strange way rather than just using the super class to handle that for us. Not really sure why, but will validate whether this works adequately. First indication looks OK.