Fix display of JavaExec executable for Rails tasks on later Gradle versions
Currently always displays `null` which is confusing. Technically it uses `JavaExec.getEffectiveExecutable` to decide the JDK version between user override, Gradle JDK and the new launcher/toolchains support, but this is private so we cannot use it. Better to just display a default if we can't determine the exact path from the command.