Reinstate Gradle Daemon JVM arg defaults that were being accidentally overridden
According to https://github.com/gradle/gradle/issues/19750 as soon as we set the `--add-opens` here, we lost the constraints on JVM size and metaspace that Gradle normally sets. Oops. This possibly explains build instability (especially on Windows) and having to constrain our number of parallel workers in order to avoid our build containers exhausting memory and dying or getting killed by the container runtime.